stc lcd english

+++ hotter/firmware/app1/Inc/filter.h
@@ -2,7 +2,7 @@
 #define FILTER_H_
 #define LOW_TEMPERATURE    19
 #define HIGH_TEMPERATURE   25
-float *low_temperature_cal(float arr[], int len);
+float *low_temperature_cal(unsigned int arr[], int len);
 void get_temp_cal(float *buf);
 #endif
